Fixup: update undoinsert.c for new undo storage interface
Patch undoinsert.c to work with the new undo log storage API.  It
has several improvements:

1.  It keeps track of transaction changes and whether the you are
at the beginning of a transaction, making life simpler for the record
insertion code.

2.  It not longer requires client code to deal with "meta data" and
"attach" records; they are replaced with new code that uses
registered buffers.

3.  It can now deal with a single WAL record that inserts data into
two different undo logs (because one ran out of space), simplifying
the client code (so it no longer has to allocate all undo space up
front).

The plan is for this to be squashed into the previous commit.

Author: Thomas Munro